requirements.txt
pydantic-settings==2.3.4
aiohttp==3.10.2
better-proxy==1.2.0
loguru==0.7.2
Pyrogram==2.0.106
TgCrypto==1.2.5
aiocache==0.12.2
pytz==2024.1
fake-useragent==1.5.1
aiohttp-socks==0.9.0
aiohttp-proxy==0.1.2
pyyaml
